+++ gutenberg/src/blocks.js
@@ -10,7 +10,7 @@ import './author-box/block.js';
 
 // CPT Post Meta
 // This works, but it's not pretty
-import './cpt-meta/block.js';
+// import './cpt-meta/block.js';
 
 // Grade
 import './grade/block.js';
